convert to topojson

+++ index.js
@@ -1,7 +1,9 @@
-const { parse } = require('csv-parse/sync'); const fs = require('fs'); const URL = require("url").URL;
+const { parse } = require('csv-parse/sync');
+const fs = require('fs');
+const URL = require("url").URL;
 const fetch = (...args) => import('node-fetch').then(({default: fetch}) => fetch(...args));
 const AdmZip = require("adm-zip");
-const shapefile = require("shapefile");
+// const shapefile = require("shapefile");
 const { spawn } = require('child_process');
 
 const stringIsAValidUrl = (s) => {
@@ -33,29 +35,54 @@ async function downloadERF() {
   fs.writeFileSync(`${__dirname}/data.json`, JSON.stringify(parsedCsv, null, 2));
 }
 
-
 function convertToGeoJson2(outputPrefix, pId) {
+  const logStream = fs.createWriteStream(`${outputPrefix}.topojson`, {flags: 'a'});
+
   return new Promise((resolve, reject) => {
-    const childProcess = spawn('ogr2ogr', [`${outputPrefix}.geojson`, `${outputPrefix}/${pId}_CEA.shp`]);
+    const childProcess = spawn('ogr2ogr', [`${outputPrefix}.geojson`, `${outputPrefix}/${pId}_CEA.shp`, '-overwrite']);
+    // const childProcess3 = spawn('geo2topo');
+
+    childProcess.stderr.on('data', (data) => {
+      console.log(`stderr: ${data}`);
+    });
+
     childProcess.on('exit', function (code, signal) {
       if (code === 0) {
-        return resolve();
+        const childProcess2 = spawn('geo2topo', [`${outputPrefix}.geojson`]);
+
+        childProcess2.stdout.pipe(logStream);
+        // childProcess2.stdout.on('data', (data) => {
+        //   console.log(`stdout: ${data}`);
+        // });
+        //
+        childProcess2.stderr.on('data', (data) => {
+          console.log(`stderr: ${data}`);
+        });
+
+        return childProcess2.on('exit', function (code2, signal) {
+          if (code2 === 0) {
+            return resolve();
+          }
+          console.log(code2, signal);
+          return reject();
+        });
       }
+      console.log(code, signal);
       return reject();
     });
   });
 }
-async function convertToGeoJson(outputPrefix, pId) {
-  const shp = await shapefile.open(`${outputPrefix}/${pId}_CEA.shp`)
-  const geoJson = await shp.read();
-
-  return new Promise((resolve, reject) => {
-    fs.writeFile(`${outputPrefix}.geojson`, JSON.stringify(geoJson.value), (err) => {
-      if (err) return reject(err);
-      return resolve();
-    });
-  });
-}
+// async function convertToGeoJson(outputPrefix, pId) {
+//   const shp = await shapefile.open(`${outputPrefix}/${pId}_CEA.shp`)
+//   const geoJson = await shp.read();
+//
+//   return new Promise((resolve, reject) => {
+//     fs.writeFile(`${outputPrefix}.geojson`, JSON.stringify(geoJson.value), (err) => {
+//       if (err) return reject(err);
+//       return resolve();
+//     });
+//   });
+// }
 
 async function processProject(project, retries=5) {
   const pId = project['Project ID'];
@@ -92,18 +119,20 @@ async function processProject(project, retries=5) {
         try {
           const zip = new AdmZip(`${outputPrefix}.zip`);
           zip.extractAllTo(`${outputPrefix}/`, true);
+          await convertToGeoJson2(outputPrefix, pId);
         } catch (e) {
+          console.log(e);
           console.log(`Retrying ${pId} (${retries - 1} retries left)`);
           return processProject(project, retries - 1);
         }
 
-        await convertToGeoJson2(outputPrefix, pId);
-
         return fs.unlink(`${outputPrefix}.zip`, (err3) => {
-          if (err3) return reject(err3);
-          console.log(`CEA ${project['Project ID']} extracted`);
-          fs.rmSync(outputPrefix, { recursive: true, force: true });
-          return resolve();
+          return fs.unlink(`${outputPrefix}.geojson`, (err4) => {
+            if (err3 || err4) return reject(err3 || err4);
+            console.log(`CEA ${project['Project ID']} extracted`);
+            fs.rmSync(outputPrefix, { recursive: true, force: true });
+            return resolve();
+          });
         });
       })
     });
